Remove a few EBNF sections for now
They were wildly outdated and it's not fun to keep them up to date.
Once the pace has settled a bit and we're closer to "completion",
I'll try to add them back in.

They're useful because the Reference doesn't contain unstable or
internal syntax, since I've reinterpeted/rephrased certain subgrammars
and since it's still kinda fun to have them.